requirements.txt
python-telegram-bot
google-generativeai
requests
beautifulsoup4
flask
gunicorn